perf: faster ShardedMap GetRand
Instead of checking random shards until we find a non-empty shard,
we generate the random number once and then we iterate the shards
until we find an non-empty shard.

This weakens the randomness of the result if there are empty shards
and the occupied shards are clustered together, but in my opinion,
this is a small price to pay when we expect sharded maps to be
filled up with millions of elements.